+++ cmd/substreams/gui.go
@@ -32,6 +32,7 @@ func init() {
 	guiCmd.Flags().Uint64("limit-processed-blocks", 10000, "Limit the number of blocks to be processed by the server, including preparing the stores, as a safeguard to prevent unexpected expensive reprocessing (0 disables the limit)")
 	guiCmd.Flags().Bool("final-blocks-only", false, "Only process blocks that have pass finality, to prevent any reorg and undo signal by staying further away from the chain HEAD")
 	guiCmd.Flags().StringSlice("debug-modules-initial-snapshot", nil, "List of 'store' modules from which to print the initial data snapshot (Unavailable in Production Mode")
+	guiCmd.Flags().StringSlice("debug-modules-output", nil, "List of modules from which to fetch the outputs, useful when debugging (unavailable in Production Mode). Defaults to the non-imported modules.")
 	guiCmd.Flags().Bool("production-mode", false, "Enable Production Mode, with high-speed parallel processing")
 	guiCmd.Flags().StringArrayP("params", "p", nil, "Set a params for parameterizable modules. Can be specified multiple times. Ex: -p module1=valA -p module2=valX&valY")
 	guiCmd.Flags().Bool("replay", false, "Replay saved session into GUI from replay.bin")
@@ -140,6 +141,11 @@ func runGui(cmd *cobra.Command, args []string) (err error) {
 		debugModulesInitialSnapshot = nil
 	}
 
+	debugModulesOutput := sflags.MustGetStringSlice(cmd, "debug-modules-output")
+	if len(debugModulesOutput) == 0 {
+		debugModulesOutput = nil
+	}
+
 	network := sflags.MustGetString(cmd, "network")
 	if network == "" {
 		network = packageBundle.Package.Network
@@ -199,6 +205,7 @@ func runGui(cmd *cobra.Command, args []string) (err error) {
 		Graph:                       packageBundle.Graph,
 		ProdMode:                    productionMode,
 		DebugModulesInitialSnapshot: debugModulesInitialSnapshot,
+		DebugModulesOutput:          debugModulesOutput,
 		Endpoint:                    endpoint,
 		OutputModule:                outputModule,
 		OverrideNetwork:             network,